started fixing master branch. noise_analysis works stand-alone now
