+++ Examples/def_config.yml
@@ -1,0 +1,48 @@
+---
+  # Some files to load in, and store data
+  Pedestal_file: '\\HEROS\dbloech\Alibava_measurements\binaryhdf5cal\pedestal_h5.h5'
+  Delay_scan: ''
+  #Charge_scan: 'C:\Users\dbloech\Desktop\MedAustron TB\calibration_100V.dat'
+  Charge_scan: '\\HEROS\dbloech\Alibava_measurements\binaryhdf5cal\calibration_m20_h5.h5'
+  Measurement_file:
+    - '\\HEROS\dbloech\Alibava_measurements\binaryhdf5cal\200V_20deg.h5'
+    - '\\HEROS\dbloech\Alibava_measurements\binaryhdf5cal\200V_10deg.h5'
+    - '\\HEROS\dbloech\Alibava_measurements\binaryhdf5cal\200V_00deg.h5'
+    - '\\HEROS\dbloech\Alibava_measurements\binaryhdf5cal\200V_m10deg.h5'
+    - '\\HEROS\dbloech\Alibava_measurements\binaryhdf5cal\200V_m20deg.h5'
+
+  Output_folder: 'C:\Users\dbloech\Desktop\Alibava'
+  Output_name: 'm20deg_cal'
+  isBinary: False # If the files provided are Alibava binaries (True) or hdf5 (False) file types
+
+  # Event analysis parameters
+  Charge_scale: True # Convert ADC to electrons
+  Processes: 1 # High numbers of processes causes huge memory overhead, only use when more than 100G are available for large files!!!
+  Plot_single_event: 1000
+  SN_cut: 6 # Minimum height of hit
+  SN_ratio: 0.5 # Ratio at which the program searches for nearby hits below the SN cut
+  SN_cluster: 5 # SN what the whole cluster must have at minimum to be considered, values great then 7 are useless
+  Noise_cut: 2 # Here the unit is ADC over median of all strips. This discriminates noisy strips.
+  Manual_mask: [38,39,40,41,42] # Manual channel masking
+  timing:
+    - 0 # Minimum timing window
+    - 250 # Maximum timing window
+  max_cluster_size: 7 # Directly affects runtime!!!
+  sensor_type: "n-in-p" # Sensor Material
+  automasking: True # Tries to find non physical or garbage hits (wrong ADC direction) and excludes them from calculation (costs moderate CPU)
+  optimize: True # Use Numba jit optimizer or not --> Warning no progress bar can be shown with this true, or may be misleading
+  additional_analysis:
+      - Langau
+      #- CCE
+      #- ChargeSharing
+
+  langau:
+    clustersize: [1,2,3] # Defines which cluster sizes should be fitted with a langau, list of values defines only these, negative values are for all clustersizes
+    numClus: [1,2,3] # Define which/how many clusters per event are taken into account for calculations
+    seed_cut_langau: True # If you want a langau from the seed cut strips only or not
+    energyCutOff: 100000 # Define where the high energy cutoff is applied in electrons
+    bins: 200 # Define how many bins the Langau plots should have
+    fitLangau: True # Fit a langau or not (sometimes the data does not represent a langau at all)
+
+
+
