+++ analysis_classes/__init__.py
@@ -1,0 +1,5 @@
+"Import classes from analysis_classes package"
+from .noise_analysis import NoiseAnalysis
+# from .main_analysis import MainAnalysis
+# from .calibration import Calibration
+# from .base_analysis import BaseAnalysis
