+++ cmd_shell.py
@@ -3,11 +3,11 @@
 import cmd
 import os
 from threading import Thread
-from analysis import *
-from utilities import *
 import matplotlib.pyplot as plt
 import pprint
 import numpy as np
+from analysis_classes import *
+from utilities import *
 
 np.set_printoptions(threshold=0, precision=2, edgeitems=2)
