+++ main.py
@@ -1,10 +1,13 @@
 # This file is the main analysis file here all processes will be started
 
-from analysis import *
 import os
 from optparse import OptionParser
 import matplotlib.pyplot as plt
 from cmd_shell import AlisysShell
+from utilities import create_dictionary, save_all_plots, save_dict
+from analysis_classes.calibration import Calibration
+from analysis_classes.noise_analysis import NoiseAnalysis
+from analysis_classes.main_analysis import MainAnalysis
 
 
 def main(args, options):
@@ -26,6 +29,41 @@ def main(args, options):
     else:
         print("No valid path parsed! Exiting")
         exit(1)
+
+def do_with_config_file(config):
+    """Starts analysis with a config file"""
+
+    # Look if a pedestal file is specified
+    if "Pedestal_file" in config:
+        noise_data = NoiseAnalysis(config["Pedestal_file"],
+                                   usejit=config.get("optimize", False),
+                                   configs=config)
+        noise_data.plot_data()
+
+    # Look if a calibration file is specified
+    if "Delay_scan" in config or "Charge_scan" in config:
+        config_data = Calibration(config.get("Delay_scan", ""),
+                                  config.get("Charge_scan", ""),
+                                  Noise_calc=noise_data,
+                                  isBinary=config.get("isBinary", False))
+        config_data.plot_data()
+
+    # Look if a pedestal file is specified
+    if "Measurement_file" in config:
+        # TODO: potential call before assignment error !!! with pedestal file
+
+        config.update({"calibration": config_data,
+                       "noise_analysis": noise_data})
+
+        event_data = MainAnalysis(config["Measurement_file"],
+                                  configs=config) # Is adictionary containing all keys and values for configuration
+        # Save the plots if specified
+        if config.get("Output_folder", "") and config.get("Output_name", ""):
+            save_all_plots(config["Output_name"], config["Output_folder"], dpi=300)
+            if config.get("Pickle_output", False):
+                save_dict(event_data.outputdata, config["Output_folder"] +
+                          "\\" + config["Output_name"] + ".dba")
+        return event_data.outputdata
 
 
 if __name__ == "__main__":
