+++ utilities.py
@@ -14,41 +14,48 @@ from six.moves import cPickle as pickle #for performance
 # COMMENT: tqdm and h5py are both missing in requirements
 from tqdm import tqdm
 import h5py
+import struct
 
-def create_dictionary(file, filepath):
-    '''Creates a dictionary with all values written in the file using yaml'''
-
-    file_string = os.path.abspath(os.getcwd() + str(filepath) + "\\" + str(file))
-    print ("Loading file: " + str(file))
+def create_dictionary(file, filepath=os.getcwd()):
+    """Import file with yaml from filepath + file and checks for type dict"""
+    file_string = os.path.join(filepath, file)
+    print("Loading file: " + str(file))
     with open(file_string, "r") as yfile:
         dic = yaml.load(yfile)
+        if not isinstance(dic, dict):
+            raise ImportError("Content of file is not a dictionary.")
         return dic
 
-def import_h5(*pathes):
+def import_h5(path):
     """
-    This functions imports hdf5 files generated by ALIBAVA.
-    If you pass several pathes, then you get list of objects back, containing the data respectively
-    :param pathes: pathes to the datafiles which should be imported
-    :return: list
+    This functions imports a hdf5 file generated by ALIBAVA.
     """
-
+    dic = h5py.File(os.path.normpath(path), 'r')
+    print("Keys: %s" % dic.keys())
+    return dic
+    # COMMENT: That should be done one level above. the func should just import 1 hdf5 file
     # Check if a list was passed
-    if isinstance(pathes[0]) == list:
-        pathes = pathes[0]
+    # if isinstance(pathes[0]) == list:
+    #     pathes = pathes[0]
 
     # First check if pathes exist and if so import
-    loaded_files = []
-    try:
-        for path in tqdm(pathes, desc= "Loading files:"):
-            if os.path.exists(os.path.normpath(path)):
-                # Now import all hdf5 files
-                loaded_files.append(h5py.File(os.path.normpath(path), 'r'))
-            else:
-                raise Exception('The path {!s} does not exist.'.format(path))
-        return loaded_files
-    except OSError as err:
-        print("Enountered an OSError: {!s}".format(err))
-        return False
+    # loaded_files = []
+    # try:
+    #     for path in tqdm(pathes, desc= "Loading files:"):
+    #         if os.path.exists(os.path.normpath(path)):
+    #             # Now import all hdf5 files
+    #             loaded_files.append(h5py.File(os.path.normpath(path), 'r'))
+    #         else:
+    #             raise Exception('The path {!s} does not exist.'.format(path))
+    #     return loaded_files
+    # except OSError as err:
+    #     print("Enountered an OSError: {!s}".format(err))
+    #     return False
+        # Now import all hdf5 files
+
+
+
+
 
 def get_xy_data(data, header=0):
     """This functions takes a list of strings, containing a header and xy data,
@@ -61,7 +68,7 @@ def get_xy_data(data, header=0):
             np2Darray[i-header] = np.array(list_data)
     return np2Darray
 
-def read_binary_Alibava(filepath):
+def read_binary(filepath):
     """Reads binary alibava files"""
 
     with open(os.path.normpath(filepath), "rb") as f:
@@ -96,27 +103,19 @@ def read_binary_Alibava(filepath):
                 if not blockheader:
                     print("Persumably end of binary file reached. Events read: {}".format(events))
                     eof = True
-
-        dict = {"header": {
-                            "noise": Noise,
-                            "pedestal": Pedestal,
-                            "Attribute:setup": None
-                            },
-                "events": {
-                            "header": Header,
-                            "signal": np.zeros((int(events),256), dtype=np.float32),
-                            "temperature": np.zeros(int(events), dtype=np.float32),
-                            "time": np.zeros(int(events), dtype=np.float32),
-                            "clock": np.zeros(int(events), dtype=np.float32)
-                            },
-                "scan": {
-                        "start": Starttime,
-                        "end": None,
-                        "value": None, # Values of cal files for example. eg. 32 pulses for a charge scan steps should be here
-                        "attribute:scan_definition": None
-                        }
-                }
-
+        # COMMENT: giving variables names like 'dict' that overwrite built-in types/functions is very bad
+        dic = {"header": {"noise": Noise,
+                          "pedestal": Pedestal,
+                          "Attribute:setup": None},
+               "events": {"header": Header,
+                          "signal": np.zeros((int(events), 256), dtype=np.float32),
+                          "temperature": np.zeros(int(events), dtype=np.float32),
+                          "time": np.zeros(int(events), dtype=np.float32),
+                          "clock": np.zeros(int(events), dtype=np.float32)},
+                "scan": {"start": Starttime,
+                         "end": None,
+                         "value": None, # Values of cal files for example. eg. 32 pulses for a charge scan steps should be here
+                         "attribute:scan_definition": None}}
         # Disect the header for the correct informations for values
         points = Header.split("|")[1].split(";")
         params = [x.strip("\x00") for x in points]
@@ -126,22 +125,22 @@ def read_binary_Alibava(filepath):
         # len(params) = 2 --> Eventfile
 
         if len(params) >= 4: # Cal file
-            dict["scan"]["value"] = np.arange(int(params[1]), int(params[2]), int(params[3]))  # aka xdata
+            dic["scan"]["value"] = np.arange(int(params[1]), int(params[2]), int(params[3]))  # aka xdata
         elif len(params) == 2: # Events file
-            dict["scan"]["value"] = np.arange(0, int(params[0]),step=1)  # aka xdata
+            dic["scan"]["value"] = np.arange(0, int(params[0]),step=1)  # aka xdata
 
         shift1 = int.from_bytes(b'0xFFFF0000',byteorder=byteorder)
         shift2 = int.from_bytes(b'0xFFFF',byteorder=byteorder)
         # decode data from data Blocks
         for i, event in enumerate(event_data):
-            dict["events"]["clock"][i] = struct.unpack("III", event[0:12])[-1]
+            dic["events"]["clock"][i] = struct.unpack("III", event[0:12])[-1]
             coded_time = struct.unpack("I", event[12:16])[0]
             #coded_time = event[12:16]
             ipart = (coded_time & shift1)>>16
             fpart = (np.sign(ipart))*(coded_time & shift2)
             time = 100*ipart+fpart
-            dict["events"]["time"][i] = time
-            dict["events"]["temperature"][i] = 0.12*struct.unpack("H", event[16:18])[0]-39.8
+            dic["events"]["time"][i] = time
+            dic["events"]["temperature"][i] = 0.12*struct.unpack("H", event[16:18])[0]-39.8
 
             # There seems to be garbage data which needs to be cut out
             padding = 18+32
@@ -149,11 +148,11 @@ def read_binary_Alibava(filepath):
             padding += 2*130+28
             part2 = list(struct.unpack("h" * 128, event[padding:padding + 2*128]))
             part1.extend(part2)
-            dict["events"]["signal"][i] = np.array(part1)
+            dic["events"]["signal"][i] = np.array(part1)
             #dict["events"]["signal"][i] =struct.unpack("H"*256, event[18:18+2*256])
             #extra = struct.unpack("d", event[18+2*256:18+2*256+4])[0]
 
-    return dict
+    return dic
 
 def read_file(filepath, binary=False):
     """Just reads a file and returns the content line by line"""
